requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
gunicorn==20.0.4
heroku==0.1.4
heroku-cli-wrapper==0.0.8
idna==2.9
numpy==1.18.3
pandas==1.0.3
Pillow==7.1.2
psycopg2==2.8.5
pygame==1.9.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
